Fixing pwm not started after os reboot

+++ inc/pwm/interfaces/pwm.hpp
@@ -12,7 +12,7 @@ class PwmIf
     virtual ~PwmIf() = default;
     virtual bool start() = 0;
     virtual bool stop() = 0;
-    virtual bool setduty(uint32_t) = 0;
+    virtual bool setduty(double) = 0;
 };
 
 } // namespace pwm
